+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased changes
 
+## 0.2.0.0
+
+- Merged the `HasElmType` and `HasElmDefinition` classes
+- Merged the `HasElmDecoder` and `HasElmDecoderDefinition` classes
+- Merged the `HasElmEncoder` and `HasElmEncoderDefinition` classes
+
 ## 0.1.0.1
 
 - Tightened generics-sop bounds to < 0.5.0